[MIG] pos_stock_available_online: Migration to 17.0
# Conflicts:
#	pos_stock_available_online/__manifest__.py
