+++ pos_stock_available_online/static/src/app/generic_components/product_card/product_card.xml
@@ -1,15 +1,13 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <templates id="template" xml:space="preserve">
-
     <t
-        t-name="ProductItem"
-        t-inherit="point_of_sale.ProductItem"
+        t-name="pos_stock_available_online.ProductCard"
+        t-inherit="point_of_sale.ProductCard"
         t-inherit-mode="extension"
-        owl="1"
     >
-        <xpath expr="//div[hasclass('product-content')]" position="inside">
-            <div t-if="env.pos.config.display_product_quantity" class="warehouse-info">
-                <div class="flex">
+        <xpath expr="//article//div[hasclass('product-content')]" position="after">
+            <div t-if="displayProductQuantity" class="warehouse-info align-items-start">
+                <div class="d-flex">
                     <div
                         t-if="warehouses.length > 1"
                         class="warehouse total"
@@ -17,20 +15,20 @@
                     >
                         <i class="fa fa-dropbox" />
                         <span
-                            t-attf-class="quantity {{total_quantity > env.pos.config.minimum_product_quantity_alert ? 'available' : 'not-available'}}"
+                            t-attf-class="quantity {{total_quantity > minimumProductQuantityAlert ? 'available' : 'not-available'}}"
                         >
                             <t t-esc="display_total_quantity" />
                         </span>
                     </div>
                 </div>
-                <div class="flex">
+                <div class="d-flex flex-column">
                     <t t-foreach="warehouses" t-as="warehouse" t-key="warehouse.id">
                         <div class="warehouse" t-att-title="warehouse.name">
                             <span class="warehouse-name">
                                 <t t-esc="warehouse.code" />
                             </span>
                             <span
-                                t-attf-class="quantity {{warehouse.quantity > env.pos.config.minimum_product_quantity_alert ? 'available' : 'not-available'}}"
+                                t-attf-class="quantity {{warehouse.quantity > minimumProductQuantityAlert ? 'available' : 'not-available'}}"
                             >
                                 <t t-esc="format_quantity(warehouse.quantity)" />
                             </span>
@@ -40,5 +38,4 @@
             </div>
         </xpath>
     </t>
-
 </templates>
