+++ pos_stock_available_online/static/src/app/screens/product_screen/product_list/product_list.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<templates id="template" xml:space="preserve">
+    <t
+        t-name="pos_stock_available_online.ProductsWidget"
+        t-inherit="point_of_sale.ProductsWidget"
+        t-inherit-mode="extension"
+    >
+        <xpath
+            expr="//div[hasclass('product-list-container')]//ProductCard"
+            position="attributes"
+        >
+            <attribute name="warehouse_info" add="product.warehouse_info" />
+            <attribute name="uom_id" add="product.uom_id" />
+        </xpath>
+    </t>
+</templates>
